🚀 Add MrklSystemPlanner extension and related classes
This commit adds a new extension project for the MrklSystemPlanner,
which is a planner that uses a natural language model to generate
plans based on user goals and available functions. The planner
communicates with the user through a series of system steps, which
consist of thoughts, actions, and observations.

The extension project defines the MrklSystemPlanner class, which
implements the IPlanner interface and interacts with the semantic
kernel and the embeddings engine. It also defines the
MrklSystemPlannerConfig class, which holds the configuration options
for the planner, such as the relevancy threshold, the maximum number
of tokens and iterations, and the lists of excluded and included
functions and skills. The SystemStep class represents a single step
in the plan, with properties for the thought, action, action input,
observation, and final answer. The class is annotated with
JsonPropertyNames for serialization.

The extension project also includes a skprompt.txt file, which
contains the prompt for the natural language model. The prompt is
based on the Mrkl document and the user goal, and it provides the
model with the available functions and their descriptions, as well as
the steps taken so far. The prompt is formatted as a JSON object and
passed to the embeddings engine as a query. The response from the
engine is parsed as a JSON object and converted to a SystemStep
object. The planner then executes the action of the step, if any, and
updates the context accordingly. The planner repeats this process
until the plan is complete or an error occurs.

This commit also renames the IterativePlanner class and namespace to
MrklSystemPlanner, to reflect the use of Mrkl systems as described in
https://arxiv.org/pdf/2205.00445.pdf. It also deletes some unused files
related to the old iterative planner, such as the chat system templates
and the WebSearchSkill class. It updates the EmbeddedResource class to
use the new namespace, and adds a new MrklSystemPlanner class that
implements the logic for creating a plan using Mrkl systems. The new
class also has a configuration object and an optional prompt override.

This commit also removes the MrklPlannerText and MrklPlannerChat classes
and their associated resources, tests, and references. They were
experimental planners that used semantic functions to create sequential
plans for text and chat completion. They are no longer needed as they
have been replaced by the more robust and flexible MrklSystemPlanner.

This commit also removes the CoreCLR-NCalc package from the
Directory.Packages.props file, as it is no longer used by any project.
It also deletes the Planning.IterativePlanner project and its
associated files, as it has been superseded by the
Planning.MrklSystemPlanner project. The SK-dotnet.sln file is updated
to reflect these changes.

This commit cleans up the codebase and reduces the number of
dependencies and projects. It also adds a new feature that allows the
user to interact with a natural language model to generate and execute
plans using available functions.
