+++ dotnet/Directory.Packages.props
@@ -6,7 +6,6 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageVersion Include="Azure.Search.Documents" Version="11.5.0-beta.2" />
-    <PackageVersion Include="CoreCLR-NCalc" Version="2.2.113" />
     <PackageVersion Include="Microsoft.Bcl.HashCode" Version="[1.1.0, )" />
     <PackageVersion Include="Microsoft.Bcl.AsyncInterfaces" Version="[7.0.0, )" />
     <PackageVersion Include="NRedisStack" Version="0.6.1" />
