--- dotnet/src/Extensions/Planning.IterativePlanner/Planning.IterativePlanner.csproj
@@ -1,42 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <Import Project="$(RepoRoot)/dotnet/nuget/nuget-package.props" />
-  <Import Project="$(RepoRoot)/dotnet/src/InternalUtilities/InternalUtilities.props" />
-
-
-  <ItemGroup>
-    <None Remove="iterative-planer-chat-system.txt" />
-    <None Remove="iterative-planer-chat-user.txt" />
-    <None Remove="iterative-planer-text.txt" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="iterative-planer-chat-user.txt">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="iterative-planer-chat-system.txt">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </EmbeddedResource>
-    <EmbeddedResource Include="iterative-planer-text.txt">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="CoreCLR-NCalc" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\Connectors\Connectors.AI.OpenAI\Connectors.AI.OpenAI.csproj" />
-    <ProjectReference Include="..\..\SemanticKernel.Abstractions\SemanticKernel.Abstractions.csproj" />
-    <ProjectReference Include="..\..\SemanticKernel\SemanticKernel.csproj" />
-    <ProjectReference Include="..\..\Skills\Skills.Web\Skills.Web.csproj" />
-  </ItemGroup>
-
-</Project>
