--- dotnet/src/Extensions/Planning.IterativePlanner/iterative-planer-chat-user.txt
@@ -1,2 +1,0 @@
-Question: {{$question}}
-{{$agentScratchPad}}
